+++ packages/chains/chains/25.ts
@@ -5,7 +5,8 @@ export default {
   "rpc": [
     "https://cronos.rpc.thirdweb.com/${THIRDWEB_API_KEY}",
     "https://evm.cronos.org",
-    "https://cronos-evm.publicnode.com"
+    "https://cronos-evm.publicnode.com",
+    "wss://cronos-evm.publicnode.com"
   ],
   "features": [
     {