+++ packages/react-native/src/evm/components/ConnectWalletDetails/WalletDetailsButton.tsx
@@ -70,6 +70,7 @@ export const WalletDetailsButton = ({
               ) : (
                 <AddressDisplay
                   variant="bodySmallSecondary"
+                  extraShort={false}
                   address={address}
                 />
               )}