+++ packages/react-native/src/evm/components/base/IconTextButton.tsx
@@ -1,18 +1,15 @@
 import { ReactNode } from "react";
 import { Theme } from "../../styles/theme";
 import BaseButton from "./BaseButton";
-import ImageSvgUri from "./ImageSvgUri";
 import Text from "./Text";
 import { StyleSheet } from "react-native";
 
 type IconTextProps = {
-  icon?:
-    | { iconUrl?: string; iconWidth?: number; iconHeight?: number }
-    | ReactNode;
+  icon?: ReactNode;
   text: string;
   padding?: keyof Theme["spacing"];
   onPress: () => void;
-} & (typeof BaseButton)["arguments"];
+} & React.ComponentProps<typeof BaseButton>;
 
 export const IconTextButton = ({
   icon,
@@ -30,15 +27,7 @@ export const IconTextButton = ({
       borderColor="border"
       {...args}
     >
-      {icon && icon.iconUrl ? (
-        <ImageSvgUri
-          imageUrl={icon.iconUrl}
-          height={icon.iconHeight}
-          width={icon.iconWidth}
-        />
-      ) : (
-        icon
-      )}
+      {icon}
       <Text variant="bodySmall" style={styles.networkText}>
         {text}
       </Text>