+++ packages/react-native/src/evm/utils/addresses.ts
@@ -7,7 +7,7 @@ import { utils } from "ethers";
  * @param extraShort If true, show the first 4 and last 3 characters
  * @returns The shortened address
  */
-export function shortenWalletAddress(str?: string, extraShort?: true) {
+export function shortenWalletAddress(str?: string, extraShort = true) {
   if (!str) {
     return str;
   }