+++ packages/react/src/evm/hooks/wallets/usePaper.ts
@@ -1,17 +1,17 @@
+import {
+  UseQueryResult,
+  useQuery,
+  useQueryClient,
+} from "@tanstack/react-query";
 import { useConnect, useWallet } from "@thirdweb-dev/react-core";
 import {
   PaperWallet,
-  walletIds,
   PaperWalletAdditionalOptions,
+  walletIds,
 } from "@thirdweb-dev/wallets";
 import { useCallback, useEffect } from "react";
-import {
-  useQuery,
-  UseQueryResult,
-  useQueryClient,
-} from "@tanstack/react-query";
 
-type PaperConfig = Omit<PaperWalletAdditionalOptions, "chain" | "chains">;
+type PaperConfig = Omit<PaperWalletAdditionalOptions, "chain">;
 
 export function usePaperWallet() {
   const connect = useConnect();
@@ -20,7 +20,7 @@ export function usePaperWallet() {
       const { paperWallet } = await import(
         "../../../wallet/wallets/paperWallet"
       );
-      return connect(paperWallet({ paperClientId: options.paperClientId }), {
+      return connect(paperWallet(options), {
         chainId: options.chainId,
         email: options.email,
       });