+++ packages/sdk/src/evm/common/gas-price.ts
@@ -25,7 +25,7 @@ export async function getDefaultGasOverrides(provider: providers.Provider) {
   );
   if (feeData.maxFeePerGas && feeData.maxPriorityFeePerGas) {
     return {
-      maxFeePerGas: feeData.maxPriorityFeePerGas,
+      maxFeePerGas: feeData.maxFeePerGas,
       maxPriorityFeePerGas: feeData.maxPriorityFeePerGas,
     };
   } else {
@@ -55,10 +55,6 @@ export async function getDynamicFeeData(
   if (chainId === Mumbai.chainId || chainId === Polygon.chainId) {
     // for polygon, get fee data from gas station
     maxPriorityFeePerGas = await getPolygonGasPriorityFee(chainId);
-    console.log(
-      "DYNAMIC: polygon fee",
-      utils.formatUnits(maxPriorityFeePerGas, "gwei"),
-    );
   } else if (eth_maxPriorityFeePerGas) {
     // prioritize fee from eth_maxPriorityFeePerGas
     maxPriorityFeePerGas = BigNumber.from(eth_maxPriorityFeePerGas);