+++ packages/sdk/src/evm/core/wallet/user-wallet.ts
@@ -13,6 +13,7 @@ import {
 } from "ethers";
 import EventEmitter from "eventemitter3";
 import invariant from "tiny-invariant";
+import { getDefaultGasOverrides } from "../../common";
 import { fetchCurrencyValue } from "../../common/currency/fetchCurrencyValue";
 import { isNativeToken } from "../../common/currency/isNativeToken";
 import { normalizePriceValue } from "../../common/currency/normalizePriceValue";
@@ -282,9 +283,33 @@ export class UserWallet {
    */
   public async sendRawTransaction(
     transactionRequest: providers.TransactionRequest,
-  ): Promise<TransactionResult> {
+  ): Promise<providers.TransactionResponse> {
     const signer = this.requireWallet();
-    const tx = await signer.sendTransaction(transactionRequest);
+    const hasGasPrice = !!transactionRequest.gasPrice;
+    const hasFeeData =
+      !!transactionRequest.maxFeePerGas &&
+      !!transactionRequest.maxPriorityFeePerGas;
+    const hasGasData = hasGasPrice || hasFeeData;
+    if (!hasGasData) {
+      // set default gas values
+      const defaultGas = await getDefaultGasOverrides(
+        this.connection.getProvider(),
+      );
+      transactionRequest.maxFeePerGas = defaultGas.maxFeePerGas;
+      transactionRequest.maxPriorityFeePerGas = defaultGas.maxPriorityFeePerGas;
+      transactionRequest.gasPrice = defaultGas.gasPrice;
+    }
+    return signer.sendTransaction(transactionRequest);
+  }
+
+  /**
+   * Execute a raw transaction to the blockchain from the connected wallet and wait for it to be mined
+   * @param transactionRequest - raw transaction data to send to the blockchain
+   */
+  public async executeRawTransaction(
+    transactionRequest: providers.TransactionRequest,
+  ): Promise<TransactionResult> {
+    const tx = await this.sendRawTransaction(transactionRequest);
     return {
       receipt: await tx.wait(),
     };